Forward propagation implementation

+++ multilayer_perceptron/activation_functions/sigmoid_activation.py
@@ -3,5 +3,5 @@ from multilayer_perceptron.activation_functions.base_exponent_activation import 
 
 class SigmoidActivation(BaseExponentActivation):
 
-    def transform(self, number: int) -> float:
+    def transform(self, number: float) -> float:
         return 1 / (1 + self.exponent_helper.calculate_exponent(-number))
